Add Spring validation and basin constraints for requests

+++ src/main/java/ua/kpi/ip22mp/trotsiuk/mmcmlr/services/RandomNumbersService.java
@@ -4,7 +4,7 @@ public interface RandomNumbersService {
 
     int[][] generateRandomIntNumbersInMatrix(int rows, int columns, int begin, int end);
 
-    double[][] generateRandomDoubleNumbersInMatrix(int rows, int columns, int begin, int end);
+    double[][] generateRandomDoubleNumbersInMatrix(int rows, int columns, double begin, double end);
 
     double[][] generateMatrixOfNormallyDistributedRandomNumbers(double mean, double stdDev, int rows, int columns);
 }
